Updated the return type for HTTP endppoint because we are returning the actual benefit object, it shoule be the updated value after injecting the env variables
