+++ src/main/java/com/kspoc/IbankingKubernetesSecretsApplication.java
@@ -4,14 +4,14 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 
-import com.kspoc.entities.DbConfig;
+import com.kspoc.entities.BenefitDetails;
 
 @SpringBootApplication
 public class IbankingKubernetesSecretsApplication {
 
 	public static void main(String[] args) {
 		ConfigurableApplicationContext context = SpringApplication.run(IbankingKubernetesSecretsApplication.class, args);
-		System.err.println(context.getBean(DbConfig.class));
+		System.err.println(context.getBean(BenefitDetails.class));
 	}
 
 }
